requirements.txt
Faker==24.2.0
psycopg2-binary==2.9.9
robotframework==7.0
robotframework-appiumlibrary==2.0.0
robotframework-browser==18.2.0
robotframework-jsonlibrary==0.5
robotframework-requests==1.0a7
robotframework-seleniumlibrary==6.1.3
python-dotenv==1.0.1